Adapt bot getting models through coach
