+++ docker/bot.Dockerfile
@@ -1,10 +1,8 @@
-FROM lappis/coach:boilerplate as coach
 FROM lappis/botrequirements:boilerplate
 
 
 COPY ./bot /bot
 COPY ./scripts /scripts
-COPY --from=coach /src_models/ /models/
 
 WORKDIR /bot
